+++ component/pom.xml
@@ -20,15 +20,15 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.wso2.extension.siddhi.execution.timeseries</groupId>
-        <artifactId>siddhi-execution-timeseries-parent</artifactId>
+        <groupId>org.wso2.extension.siddhi.execution.extrema</groupId>
+        <artifactId>siddhi-execution-extrema-parent</artifactId>
         <version>3.1.3-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>siddhi-execution-timeseries</artifactId>
+    <artifactId>siddhi-execution-extrema</artifactId>
     <packaging>bundle</packaging>
-    <name>WSO2 Siddhi Execution Extension - Time Series</name>
+    <name>WSO2 Siddhi Execution Extension - Extrema</name>
     <url>http://wso2.org</url>
 
     <dependencies>
@@ -83,8 +83,8 @@
                         </Private-Package>
                         <Export-Package>
                             !Jama.*,
-                            org.wso2.extension.siddhi.execution.timeseries,
-                            org.wso2.extension.siddhi.execution.timeseries.*
+                            org.wso2.extension.siddhi.execution.extrema,
+                            org.wso2.extension.siddhi.execution.extrema.*
                         </Export-Package>
                         <Import-Package>
                             *;resolution:=optional
