+++ component/src/main/java/org/wso2/extension/siddhi/execution/extrema/LengthTimeLinearRegressionOutlierStreamProcessor.java
@@ -15,11 +15,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.extension.siddhi.execution.timeseries;
+package org.wso2.extension.siddhi.execution.extrema;
 
-import org.wso2.extension.siddhi.execution.timeseries.linreg.LengthTimeMultipleLinearRegression;
-import org.wso2.extension.siddhi.execution.timeseries.linreg.LengthTimeRegressionCalculator;
-import org.wso2.extension.siddhi.execution.timeseries.linreg.LengthTimeSimpleLinearRegressionCalculator;
+import org.wso2.extension.siddhi.execution.extrema.linreg.LengthTimeSimpleLinearRegressionCalculator;
 import org.wso2.siddhi.core.config.ExecutionPlanContext;
 import org.wso2.siddhi.core.event.ComplexEventChunk;
 import org.wso2.siddhi.core.event.stream.StreamEvent;
@@ -30,38 +28,39 @@ import org.wso2.siddhi.core.executor.ConstantExpressionExecutor;
 import org.wso2.siddhi.core.executor.ExpressionExecutor;
 import org.wso2.siddhi.core.query.processor.Processor;
 import org.wso2.siddhi.core.query.processor.stream.StreamProcessor;
+import org.wso2.extension.siddhi.execution.extrema.linreg.LengthTimeRegressionCalculator;
 import org.wso2.siddhi.query.api.definition.AbstractDefinition;
 import org.wso2.siddhi.query.api.definition.Attribute;
 import java.util.ArrayList;
 import java.util.List;
 
 /*
- * Sample Query1 (time window, length window, y, x):
- * from InputStream#timeseries:lengthTimeOutlier(20 min, 20, y, x)
+ * Sample Query1 (time window, length window, range, y, x):
+ * from InputStream#extrema:lengthTimeOutlier(20 min, 20, 1, y, x)
  * select *
  * insert into OutputStream;
  *
- * Sample Query2 (time window, length window, calculation interval, confidence interval, y, x1, x2):
- * from InputStream#timeseries:lengthTimeOutlier(20 min, 20, 2, 0.9, y, x1, x2)
+ * Sample Query2 (time window, length window, range, calculation interval, confidence interval, y, x):
+ * from InputStream#extrema:lengthTimeOutlier(20 min, 20, 1, 2, 0.9, y, x)
  * select *
  * insert into OutputStream;
  *
- * This class performs linear regression.
+ * This class detects outliers based on simple linear regression.
  * Number of data points could be constrained using both time and length windows.
  */
-public class LengthTimeLinearRegressionStreamProcessor extends StreamProcessor {
+public class LengthTimeLinearRegressionOutlierStreamProcessor extends StreamProcessor {
     private int paramCount; // Number of x variables +1
     private long duration; // Time window to consider for regression calculation
     private int calcInterval = 1; // The frequency of regression calculation
-    private double ci = 0.95; // Confidence Interval
-    // simple linear regression
+    private double ci = 0.95; // Confidence Interval simple linear regression
     private LengthTimeRegressionCalculator regressionCalculator = null;
     private int yParameterPosition;
+    private Object[] coefficients;
     private static final int SIMPLE_LINREG_INPUT_PARAM_COUNT = 2; //Number of input parameters in
                                                                   // simple linear regression
 
     /**
-     * The init method of the LinearRegressionStreamProcessor,
+     * The init method of the LinearRegressionOutlierStreamProcessor,
      * this method will be called before other methods
      *
      * @param inputDefinition the incoming stream definition
@@ -73,8 +72,9 @@ public class LengthTimeLinearRegressionStreamProcessor extends StreamProcessor {
     protected List<Attribute> init(AbstractDefinition inputDefinition,
                                    ExpressionExecutor[] attributeExpressionExecutors,
                                    ExecutionPlanContext executionPlanContext) {
-        paramCount = attributeExpressionLength - 2; // First two events are time, length windows.
-        yParameterPosition = 2;
+        paramCount = attributeExpressionLength - 3; // First three events are time window, length
+                                                    // window and range
+        yParameterPosition = 3;
         // Capture duration
         if (attributeExpressionExecutors[0] instanceof ConstantExpressionExecutor) {
             if (attributeExpressionExecutors[0].getReturnType() == Attribute.Type.INT) {
@@ -95,41 +95,42 @@ public class LengthTimeLinearRegressionStreamProcessor extends StreamProcessor {
         int batchSize; // Maximum # of events, used for regression calculation
         if (attributeExpressionExecutors[1] instanceof ConstantExpressionExecutor) {
             if (attributeExpressionExecutors[1].getReturnType() == Attribute.Type.INT) {
-                batchSize = (Integer)
-                        ((ConstantExpressionExecutor) attributeExpressionExecutors[1]).getValue();
+                batchSize = (Integer) ((ConstantExpressionExecutor)
+                        attributeExpressionExecutors[1]).getValue();
             } else {
-                throw new ExecutionPlanCreationException("Size parameter should be int, but found "
-                        + attributeExpressionExecutors[1].getReturnType());
+                throw new ExecutionPlanCreationException
+                        ("Size parameter should be int, but found "
+                                + attributeExpressionExecutors[1].getReturnType());
             }
         } else {
             throw new ExecutionPlanCreationException("Size parameter must be a constant");
         }
         // Capture calculation interval and ci if provided by user
         // Default values would be used otherwise
-        if (attributeExpressionExecutors[2] instanceof ConstantExpressionExecutor) {
+        if (attributeExpressionExecutors[3] instanceof ConstantExpressionExecutor) {
             paramCount = paramCount - 2; // When calcInterval and ci are given by user,
                                          // parameter count must exclude those two as well
-            yParameterPosition = 4;
-            if (attributeExpressionExecutors[2].getReturnType() == Attribute.Type.INT) {
-                calcInterval = (Integer)
-                        ((ConstantExpressionExecutor) attributeExpressionExecutors[2]).getValue();
+            yParameterPosition = 5;
+            if (attributeExpressionExecutors[3].getReturnType() == Attribute.Type.INT) {
+                calcInterval = (Integer) ((ConstantExpressionExecutor)
+                        attributeExpressionExecutors[3]).getValue();
             } else {
-                throw new ExecutionPlanCreationException("Calculation interval should be " +
-                                                         "int, but found "
-                        + attributeExpressionExecutors[2].getReturnType());
+                throw new ExecutionPlanCreationException
+                        ("Calculation interval should be int, but found "
+                                + attributeExpressionExecutors[3].getReturnType());
             }
-            if (attributeExpressionExecutors[3] instanceof ConstantExpressionExecutor) {
-                if (attributeExpressionExecutors[3].getReturnType() == Attribute.Type.DOUBLE) {
+            if (attributeExpressionExecutors[4] instanceof ConstantExpressionExecutor) {
+                if (attributeExpressionExecutors[4].getReturnType() == Attribute.Type.DOUBLE) {
                     ci = (Double) ((ConstantExpressionExecutor)
-                            attributeExpressionExecutors[3]).getValue();
+                            attributeExpressionExecutors[4]).getValue();
                     if (!(0 <= ci && ci <= 1)) {
-                        throw new ExecutionPlanCreationException(
-                                "Confidence interval should be a value between 0 and 1");
+                        throw new ExecutionPlanCreationException
+                                ("Confidence interval should be a value between 0 and 1");
                     }
                 } else {
-                    throw new ExecutionPlanCreationException("Confidence interval should " +
-                            "be double, but found "
-                            + attributeExpressionExecutors[3].getReturnType());
+                    throw new ExecutionPlanCreationException
+                            ("Confidence interval should be double, but found "
+                                    + attributeExpressionExecutors[4].getReturnType());
                 }
             } else {
                 throw new ExecutionPlanCreationException("Confidence interval must be a constant");
@@ -137,20 +138,21 @@ public class LengthTimeLinearRegressionStreamProcessor extends StreamProcessor {
         }
         // Pick the appropriate regression calculator
         if (paramCount > SIMPLE_LINREG_INPUT_PARAM_COUNT) {
-            regressionCalculator = new LengthTimeMultipleLinearRegression(paramCount,
-                    duration, batchSize, calcInterval, ci);
+            throw new ExecutionPlanCreationException("Outlier Function is available only for " +
+                    "simple linear regression");
         } else {
             regressionCalculator = new LengthTimeSimpleLinearRegressionCalculator(paramCount,
                     duration, batchSize, calcInterval, ci);
         }
-        // Add attributes for standard error and all beta values
+        // Create attributes for standard error and all beta values and the outlier result
         String betaVal;
-        List<Attribute> attributes = new ArrayList<Attribute>(paramCount + 1);
+        List<Attribute> attributes = new ArrayList<Attribute>(paramCount + 2);
         attributes.add(new Attribute("stderr", Attribute.Type.DOUBLE));
         for (int itr = 0; itr < paramCount; itr++) {
             betaVal = "beta" + itr;
             attributes.add(new Attribute(betaVal, Attribute.Type.DOUBLE));
         }
+        attributes.add(new Attribute("outlier", Attribute.Type.BOOL));
         return attributes;
     }
 
@@ -171,16 +173,46 @@ public class LengthTimeLinearRegressionStreamProcessor extends StreamProcessor {
                 StreamEvent streamEvent = streamEventChunk.next();
                 long currentTime = executionPlanContext.getTimestampGenerator().currentTime();
                 long eventExpiryTime = currentTime + duration;
+                Boolean result = false; // Becomes true if its an outlier
                 Object[] inputData = new Object[paramCount];
+                //Capture range: number of standard deviations from the regression equation
+                final int rangePosition;
+                rangePosition = 2;
+                double range = ((Number)
+                        attributeExpressionExecutors[rangePosition].execute(streamEvent)).doubleValue();
                 for (int i = yParameterPosition; i < attributeExpressionLength; i++) {
-                    inputData[i - yParameterPosition] = attributeExpressionExecutors[i].execute(streamEvent);
+                    inputData[i - yParameterPosition] =
+                            attributeExpressionExecutors[i].execute(streamEvent);
                 }
-                Object[] outputData = regressionCalculator.calculateLinearRegression(inputData,
+                if (coefficients != null) {
+                    // Get the current Y value and X value
+                    double nextY = ((Number) inputData[0]).doubleValue();
+                    double nextX = ((Number) inputData[1]).doubleValue();
+                    // Get the last computed regression coefficients
+                    double stdError = ((Number) coefficients[0]).doubleValue();
+                    double beta0 = ((Number) coefficients[1]).doubleValue();
+                    double beta1 = ((Number) coefficients[2]).doubleValue();
+                    // Forecast Y based on current coefficients and next X value
+                    double forecastY = beta0 + beta1 * nextX;
+                    // Create the normal range based on user provided range parameter and current
+                    // std error
+                    double upLimit = forecastY + range * stdError;
+                    double downLimit = forecastY - range * stdError;
+                    // Check whether next Y value is an outlier based on the next X value and the
+                    // current regression equation
+                    if (nextY < downLimit || nextY > upLimit) {
+                        result = true;
+                    }
+                }
+                // Perform regression including X and Y of current event
+                coefficients = regressionCalculator.calculateLinearRegression(inputData,
                         eventExpiryTime);
-                // Skip processing if user has specified calculation interval
-                if (outputData == null) {
+                if (coefficients == null) {
                     streamEventChunk.remove();
                 } else {
+                    Object[] outputData = new Object[coefficients.length + 1];
+                    System.arraycopy(coefficients, 0, outputData, 0, coefficients.length);
+                    outputData[coefficients.length] = result;
                     complexEventPopulater.populateComplexEvent(streamEvent, outputData);
                 }
             }
