+++ component/src/main/java/org/wso2/extension/siddhi/execution/extrema/LinearRegressionForecastStreamProcessor.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.timeseries;
+package org.wso2.extension.siddhi.execution.extrema;
 
 import org.wso2.siddhi.core.config.ExecutionPlanContext;
 import org.wso2.siddhi.core.event.ComplexEvent;
@@ -29,41 +29,61 @@ import org.wso2.siddhi.core.executor.ConstantExpressionExecutor;
 import org.wso2.siddhi.core.executor.ExpressionExecutor;
 import org.wso2.siddhi.core.query.processor.Processor;
 import org.wso2.siddhi.core.query.processor.stream.StreamProcessor;
-import org.wso2.extension.siddhi.execution.timeseries.linreg.MultipleLinearRegressionCalculator;
-import org.wso2.extension.siddhi.execution.timeseries.linreg.RegressionCalculator;
-import org.wso2.extension.siddhi.execution.timeseries.linreg.SimpleLinearRegressionCalculator;
+import org.wso2.extension.siddhi.execution.extrema.linreg.RegressionCalculator;
+import org.wso2.extension.siddhi.execution.extrema.linreg.SimpleLinearRegressionCalculator;
 import org.wso2.siddhi.query.api.definition.AbstractDefinition;
 import org.wso2.siddhi.query.api.definition.Attribute;
-
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * The methods supported by this function are
- * timeseries:regress(int/long/float/double y, int/long/float/double x1, int/long/float/double x2 ...)
- * and
- * timeseries:regress(int calcInterval, int batchSize, double confidenceInterval, int/long/float/double y, int/long/float/double x1, int/long/float/double x2 ...)
- */
-
-public class LinearRegressionStreamProcessor extends StreamProcessor {
+public class LinearRegressionForecastStreamProcessor extends StreamProcessor {
 
     private int paramCount = 0;                                         // Number of x variables +1
     private int calcInterval = 1;                                       // The frequency of regression calculation
     private int batchSize = 1000000000;                                 // Maximum # of events, used for regression calculation
     private double ci = 0.95;                                           // Confidence Interval
-    private final int SIMPLE_LINREG_INPUT_PARAM_COUNT = 2;              // Number of Input parameters in a simple linear regression
+    private final int SIMPLE_LINREG_INPUT_PARAM_COUNT = 2;              // Number of Input parameters in a simple linear forecast
     private RegressionCalculator regressionCalculator = null;
     private int paramPosition = 0;
 
+    @Override
+    protected void process(ComplexEventChunk<StreamEvent> streamEventChunk, Processor nextProcessor, StreamEventCloner streamEventCloner, ComplexEventPopulater complexEventPopulater) {
+        synchronized (this) {
+            while (streamEventChunk.hasNext()) {
+                ComplexEvent complexEvent = streamEventChunk.next();
+
+                Object[] inputData = new Object[attributeExpressionLength - paramPosition];
+
+                // Obtain x value that user wants to use to forecast Y
+                double xDash = ((Number) attributeExpressionExecutors[paramPosition - 1].execute(complexEvent)).doubleValue();
+                for (int i = paramPosition; i < attributeExpressionLength; i++) {
+                    inputData[i - paramPosition] = attributeExpressionExecutors[i].execute(complexEvent);
+                }
+                Object[] coefficients = regressionCalculator.calculateLinearRegression(inputData);
+
+                if (coefficients == null) {
+                    streamEventChunk.remove();
+                } else {
+                    Object[] outputData = new Object[coefficients.length + 1];
+                    System.arraycopy(coefficients, 0, outputData, 0, coefficients.length);
+
+                    // Calculating forecast Y based on regression equation and given x
+                    outputData[coefficients.length] = ((Number) coefficients[coefficients.length - 2]).doubleValue() + ((Number) coefficients[coefficients.length - 1]).doubleValue() * xDash;
+                    complexEventPopulater.populateComplexEvent(complexEvent, outputData);
+                }
+            }
+        }
+        nextProcessor.process(streamEventChunk);
+    }
 
     @Override
     protected List<Attribute> init(AbstractDefinition inputDefinition, ExpressionExecutor[] attributeExpressionExecutors, ExecutionPlanContext executionPlanContext) {
         paramCount = attributeExpressionLength;
 
-        // Capture constant inputs
-        if (attributeExpressionExecutors[0] instanceof ConstantExpressionExecutor) {
-            paramCount = paramCount - 3;
-            paramPosition = 3;
+        // Capture Constant inputs
+        if (attributeExpressionExecutors[1] instanceof ConstantExpressionExecutor) {
+            paramCount = paramCount - 4;
+            paramPosition = 4;
             try {
                 calcInterval = ((Integer) attributeExpressionExecutors[0].execute(null));
                 batchSize = ((Integer) attributeExpressionExecutors[1].execute(null));
@@ -82,46 +102,22 @@ public class LinearRegressionStreamProcessor extends StreamProcessor {
 
         // Pick the appropriate regression calculator
         if (paramCount > SIMPLE_LINREG_INPUT_PARAM_COUNT) {
-            regressionCalculator = new MultipleLinearRegressionCalculator(paramCount, calcInterval, batchSize, ci);
+            throw new ExecutionPlanCreationException("Forecast Function is available only for simple linear regression");
         } else {
             regressionCalculator = new SimpleLinearRegressionCalculator(paramCount, calcInterval, batchSize, ci);
         }
 
-
-        // Add attributes for standard error and all beta values
+        // Create attributes for standard error and all beta values and the Forecast Y value
         String betaVal;
-        ArrayList<Attribute> attributes = new ArrayList<Attribute>(paramCount);
+        ArrayList<Attribute> attributes = new ArrayList<Attribute>(paramCount + 1);
         attributes.add(new Attribute("stderr", Attribute.Type.DOUBLE));
 
         for (int itr = 0; itr < paramCount; itr++) {
             betaVal = "beta" + itr;
             attributes.add(new Attribute(betaVal, Attribute.Type.DOUBLE));
         }
+        attributes.add(new Attribute("forecastY", Attribute.Type.DOUBLE));
         return attributes;
-    }
-
-    @Override
-    protected void process(ComplexEventChunk<StreamEvent> streamEventChunk, Processor nextProcessor, StreamEventCloner streamEventCloner, ComplexEventPopulater complexEventPopulater) {
-        synchronized (this) {
-            while (streamEventChunk.hasNext()) {
-                ComplexEvent complexEvent = streamEventChunk.next();
-
-                Object[] inputData = new Object[attributeExpressionLength - paramPosition];
-                for (int i = paramPosition; i < attributeExpressionLength; i++) {
-                    inputData[i - paramPosition] = attributeExpressionExecutors[i].execute(complexEvent);
-                }
-                Object[] outputData = regressionCalculator.calculateLinearRegression(inputData);
-
-                // Skip processing if user has specified calculation interval
-                if (outputData == null) {
-                    streamEventChunk.remove();
-                } else {
-                    complexEventPopulater.populateComplexEvent(complexEvent, outputData);
-                }
-            }
-        }
-        nextProcessor.process(streamEventChunk);
-
     }
 
     @Override
@@ -143,5 +139,4 @@ public class LinearRegressionStreamProcessor extends StreamProcessor {
     public void restoreState(Object[] state) {
 
     }
-
 }
