+++ component/src/main/java/org/wso2/extension/siddhi/execution/extrema/LinearRegressionOutlierStreamProcessor.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.timeseries;
+package org.wso2.extension.siddhi.execution.extrema;
 
 import org.wso2.siddhi.core.config.ExecutionPlanContext;
 import org.wso2.siddhi.core.event.ComplexEvent;
@@ -29,14 +29,15 @@ import org.wso2.siddhi.core.executor.ConstantExpressionExecutor;
 import org.wso2.siddhi.core.executor.ExpressionExecutor;
 import org.wso2.siddhi.core.query.processor.Processor;
 import org.wso2.siddhi.core.query.processor.stream.StreamProcessor;
-import org.wso2.extension.siddhi.execution.timeseries.linreg.RegressionCalculator;
-import org.wso2.extension.siddhi.execution.timeseries.linreg.SimpleLinearRegressionCalculator;
+import org.wso2.extension.siddhi.execution.extrema.linreg.RegressionCalculator;
+import org.wso2.extension.siddhi.execution.extrema.linreg.SimpleLinearRegressionCalculator;
 import org.wso2.siddhi.query.api.definition.AbstractDefinition;
 import org.wso2.siddhi.query.api.definition.Attribute;
+
 import java.util.ArrayList;
 import java.util.List;
 
-public class LinearRegressionForecastStreamProcessor extends StreamProcessor {
+public class LinearRegressionOutlierStreamProcessor extends StreamProcessor {
 
     private int paramCount = 0;                                         // Number of x variables +1
     private int calcInterval = 1;                                       // The frequency of regression calculation
@@ -44,31 +45,54 @@ public class LinearRegressionForecastStreamProcessor extends StreamProcessor {
     private double ci = 0.95;                                           // Confidence Interval
     private final int SIMPLE_LINREG_INPUT_PARAM_COUNT = 2;              // Number of Input parameters in a simple linear forecast
     private RegressionCalculator regressionCalculator = null;
-    private int paramPosition = 0;
+    private int paramPosition = 1;
+    private Object[] coefficients;
 
     @Override
     protected void process(ComplexEventChunk<StreamEvent> streamEventChunk, Processor nextProcessor, StreamEventCloner streamEventCloner, ComplexEventPopulater complexEventPopulater) {
         synchronized (this) {
             while (streamEventChunk.hasNext()) {
                 ComplexEvent complexEvent = streamEventChunk.next();
+                Boolean result = false; // Becomes true if its an outlier
 
                 Object[] inputData = new Object[attributeExpressionLength - paramPosition];
+                double range = ((Number) attributeExpressionExecutors[paramPosition - 1].execute(complexEvent)).doubleValue();
 
-                // Obtain x value that user wants to use to forecast Y
-                double xDash = ((Number) attributeExpressionExecutors[paramPosition - 1].execute(complexEvent)).doubleValue();
                 for (int i = paramPosition; i < attributeExpressionLength; i++) {
                     inputData[i - paramPosition] = attributeExpressionExecutors[i].execute(complexEvent);
                 }
-                Object[] coefficients = regressionCalculator.calculateLinearRegression(inputData);
+
+                if (coefficients != null) {
+                    // Get the current Y value and X value
+                    double nextY = ((Number) inputData[0]).doubleValue();
+                    double nextX = ((Number) inputData[1]).doubleValue();
+
+                    // Get the last computed regression coefficients
+                    double stdError = ((Number) coefficients[0]).doubleValue();
+                    double beta0 = ((Number) coefficients[1]).doubleValue();
+                    double beta1 = ((Number) coefficients[2]).doubleValue();
+
+                    // Forecast Y based on current coefficients and next X value
+                    double forecastY = beta0 + beta1 * nextX;
+
+                    // Create the normal range based on user provided range parameter and current std error
+                    double upLimit = forecastY + range * stdError;
+                    double downLimit = forecastY - range * stdError;
+
+                    // Check whether next Y value is an outlier based on the next X value and the current regression equation
+                    if (nextY < downLimit || nextY > upLimit) {
+                        result = true;
+                    }
+                }
+                // Perform regression including X and Y of current event
+                coefficients = regressionCalculator.calculateLinearRegression(inputData);
 
                 if (coefficients == null) {
                     streamEventChunk.remove();
                 } else {
                     Object[] outputData = new Object[coefficients.length + 1];
                     System.arraycopy(coefficients, 0, outputData, 0, coefficients.length);
-
-                    // Calculating forecast Y based on regression equation and given x
-                    outputData[coefficients.length] = ((Number) coefficients[coefficients.length - 2]).doubleValue() + ((Number) coefficients[coefficients.length - 1]).doubleValue() * xDash;
+                    outputData[coefficients.length] = result;
                     complexEventPopulater.populateComplexEvent(complexEvent, outputData);
                 }
             }
@@ -78,11 +102,10 @@ public class LinearRegressionForecastStreamProcessor extends StreamProcessor {
 
     @Override
     protected List<Attribute> init(AbstractDefinition inputDefinition, ExpressionExecutor[] attributeExpressionExecutors, ExecutionPlanContext executionPlanContext) {
-        paramCount = attributeExpressionLength;
+        paramCount = attributeExpressionLength - 1;
 
-        // Capture Constant inputs
         if (attributeExpressionExecutors[1] instanceof ConstantExpressionExecutor) {
-            paramCount = paramCount - 4;
+            paramCount = paramCount - 3;
             paramPosition = 4;
             try {
                 calcInterval = ((Integer) attributeExpressionExecutors[0].execute(null));
@@ -102,12 +125,12 @@ public class LinearRegressionForecastStreamProcessor extends StreamProcessor {
 
         // Pick the appropriate regression calculator
         if (paramCount > SIMPLE_LINREG_INPUT_PARAM_COUNT) {
-            throw new ExecutionPlanCreationException("Forecast Function is available only for simple linear regression");
+            throw new ExecutionPlanCreationException("Outlier Function is available only for simple linear regression");
         } else {
             regressionCalculator = new SimpleLinearRegressionCalculator(paramCount, calcInterval, batchSize, ci);
         }
 
-        // Create attributes for standard error and all beta values and the Forecast Y value
+        // Create attributes for standard error and all beta values and the outlier result
         String betaVal;
         ArrayList<Attribute> attributes = new ArrayList<Attribute>(paramCount + 1);
         attributes.add(new Attribute("stderr", Attribute.Type.DOUBLE));
@@ -116,7 +139,7 @@ public class LinearRegressionForecastStreamProcessor extends StreamProcessor {
             betaVal = "beta" + itr;
             attributes.add(new Attribute(betaVal, Attribute.Type.DOUBLE));
         }
-        attributes.add(new Attribute("forecastY", Attribute.Type.DOUBLE));
+        attributes.add(new Attribute("outlier", Attribute.Type.BOOL));
         return attributes;
     }
