+++ component/src/main/java/org/wso2/extension/siddhi/execution/extrema/extrema/util/ExtremaCalculator.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.timeseries.extrema.util;
+package org.wso2.extension.siddhi.execution.extrema.extrema.util;
 
 
 import java.util.Iterator;
