+++ component/src/main/java/org/wso2/extension/siddhi/execution/extrema/linreg/LengthTimeRegressionCalculator.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.extension.siddhi.execution.timeseries.linreg;
+package org.wso2.extension.siddhi.execution.extrema.linreg;
 
 import java.util.PriorityQueue;
 import java.util.Queue;
