+++ component/src/main/resources/extrema.siddhiext
@@ -1,0 +1,25 @@
+#
+# Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+#
+# WSO2 Inc. licenses this file to you under the Apache License,
+# Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+lengthTimeRegress=org.wso2.extension.siddhi.execution.extrema.LengthTimeLinearRegressionStreamProcessor
+regress=org.wso2.extension.siddhi.execution.extrema.LinearRegressionStreamProcessor
+lengthTimeForecast=org.wso2.extension.siddhi.execution.extrema.LengthTimeLinearRegressionForecastStreamProcessor
+forecast=org.wso2.extension.siddhi.execution.extrema.LinearRegressionForecastStreamProcessor
+lengthTimeOutlier=org.wso2.extension.siddhi.execution.extrema.LengthTimeLinearRegressionOutlierStreamProcessor
+outlier=org.wso2.extension.siddhi.execution.extrema.LinearRegressionOutlierStreamProcessor
+kalmanMinMax=org.wso2.extension.siddhi.execution.extrema.extrema.KalmanMinMaxStreamProcessor
+kernelMinMax=org.wso2.extension.siddhi.execution.extrema.extrema.KernelMinMaxStreamProcessor
