+++ component/src/test/java/org/wso2/extension/siddhi/execution/extrema/KernelExtensionTestCase.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.timeseries;
+package org.wso2.extension.siddhi.execution.extrema;
 
 import junit.framework.Assert;
 import org.apache.log4j.Logger;
@@ -47,7 +47,7 @@ public class KernelExtensionTestCase {
         SiddhiManager siddhiManager = new SiddhiManager();
 
         String inStreamDefinition = "define stream inputStream (price double);";
-        String query = ("@info(name = 'query1') from inputStream#timeseries:kernelMinMax(price, 4, 17, 'max') " +
+        String query = ("@info(name = 'query1') from inputStream#extrema:kernelMinMax(price, 4, 17, 'max') " +
                 "select *" +
                 "insert into outputStream;");
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(inStreamDefinition + query);
@@ -206,7 +206,7 @@ public class KernelExtensionTestCase {
         SiddhiManager siddhiManager = new SiddhiManager();
 
         String inStreamDefinition = "define stream inputStream (id int, price double);";
-        String query = ("@info(name = 'query1') from inputStream#timeseries:kernelMinMax(price, 3 , 16, 'minmax') " +
+        String query = ("@info(name = 'query1') from inputStream#extrema:kernelMinMax(price, 3 , 16, 'minmax') " +
                 "select price, extremaType , id " +
                 "insert into outputStream;");
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(inStreamDefinition + query);
@@ -394,7 +394,7 @@ public class KernelExtensionTestCase {
         SiddhiManager siddhiManager = new SiddhiManager();
 
         String inStreamDefinition = "define stream inputStream (price double);";
-        String query = ("@info(name = 'query1') from inputStream#timeseries:kernelMinMax(price, 4, 16, 'min') " +
+        String query = ("@info(name = 'query1') from inputStream#extrema:kernelMinMax(price, 4, 16, 'min') " +
                 "select *" +
                 "insert into outputStream;");
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(inStreamDefinition + query);
