+++ component/src/test/java/org/wso2/extension/siddhi/execution/extrema/LengthTimeLinearRegressionForecastTestcase.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.timeseries;
+package org.wso2.extension.siddhi.execution.extrema;
 
 import org.apache.log4j.Logger;
 import org.junit.Assert;
@@ -49,7 +49,7 @@ public class LengthTimeLinearRegressionForecastTestcase {
         String inputStream = "define stream InputStream (y double, symbol string, x double);";
 
         // Limit number of events based on length window (query):
-        String executionPlan = ("@info(name = 'query1') from InputStream#timeseries:lengthTimeForecast(20 min, 20, x+2, 2, 0.95, y, x) "
+        String executionPlan = ("@info(name = 'query1') from InputStream#extrema:lengthTimeForecast(20 min, 20, x+2, 2, 0.95, y, x) "
                 + "select * "
                 + "insert into OutputStream;");
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(inputStream + executionPlan);
@@ -137,7 +137,7 @@ public class LengthTimeLinearRegressionForecastTestcase {
         String inputStream = "define stream InputStream (y double, symbol string, x double);";
 
         // Limit number of events based on time window (query):
-        String executionPlan = ("@info(name = 'query1') from InputStream#timeseries:lengthTimeForecast(200, 100000, x+2, 2, 0.95, y, x) "
+        String executionPlan = ("@info(name = 'query1') from InputStream#extrema:lengthTimeForecast(200, 100000, x+2, 2, 0.95, y, x) "
                 + "select * insert into OutputStream;");
         ExecutionPlanRuntime executionPlanRuntime = siddhiManager.createExecutionPlanRuntime(inputStream + executionPlan);
