+++ feature/pom.xml
@@ -19,22 +19,22 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.wso2.extension.siddhi.execution.timeseries</groupId>
-        <artifactId>siddhi-execution-timeseries-parent</artifactId>
+        <groupId>org.wso2.extension.siddhi.execution.extrema</groupId>
+        <artifactId>siddhi-execution-extrema-parent</artifactId>
         <version>3.1.3-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>siddhi-execution-timeseries-feature</artifactId>
+    <artifactId>siddhi-execution-extrema-feature</artifactId>
     <packaging>pom</packaging>
-    <name>WSO2 Siddhi Execution Extension - Time Series Feature</name>
+    <name>WSO2 Siddhi Execution Extension - Extrema Feature</name>
     <url>http://wso2.org</url>
-    <description>This feature contains Siddhi extension feature for Time Series.</description>
+    <description>This feature contains Siddhi extension feature for Extrema.</description>
 
     <dependencies>
         <dependency>
-            <groupId>org.wso2.extension.siddhi.execution.timeseries</groupId>
-            <artifactId>siddhi-execution-timeseries</artifactId>
+            <groupId>org.wso2.extension.siddhi.execution.extrema</groupId>
+            <artifactId>siddhi-execution-extrema</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
@@ -53,7 +53,7 @@
                             <goal>p2-feature-gen</goal>
                         </goals>
                         <configuration>
-                            <id>siddhi-execution-timeseries</id>
+                            <id>siddhi-execution-extrema</id>
                             <propertiesFile>feature.properties</propertiesFile>
                             <adviceFile>
                                 <properties>
@@ -63,7 +63,7 @@
                             </adviceFile>
                             <bundles>
                                 <bundleDef>
-                                    org.wso2.extension.siddhi.execution.timeseries:siddhi-execution-timeseries
+                                    org.wso2.extension.siddhi.execution.extrema:siddhi-execution-extrema
                                 </bundleDef>
                             </bundles>
                         </configuration>
