+++ pom.xml
@@ -26,11 +26,11 @@
         <version>1</version>
     </parent>
 
-    <groupId>org.wso2.extension.siddhi.execution.timeseries</groupId>
-    <artifactId>siddhi-execution-timeseries-parent</artifactId>
+    <groupId>org.wso2.extension.siddhi.execution.extrema</groupId>
+    <artifactId>siddhi-execution-extrema-parent</artifactId>
     <version>3.1.3-SNAPSHOT</version>
     <packaging>pom</packaging>
-    <name>WSO2 Siddhi Execution Extension - Time Series Aggregator Pom</name>
+    <name>WSO2 Siddhi Execution Extension - Extrema Aggregator Pom</name>
     <url>http://wso2.org</url>
 
     <profiles>
@@ -55,9 +55,9 @@
     </properties>
 
     <scm>
-        <connection>scm:git:https://github.com/wso2-extensions/siddhi-execution-timeseries.git</connection>
-        <url>https://github.com/wso2-extensions/siddhi-execution-timeseries.git</url>
-        <developerConnection>scm:git:https://github.com/wso2-extensions/siddhi-execution-timeseries.git
+        <connection>scm:git:https://github.com/wso2-extensions/siddhi-execution-extrema.git</connection>
+        <url>https://github.com/wso2-extensions/siddhi-execution-extrema.git</url>
+        <developerConnection>scm:git:https://github.com/wso2-extensions/siddhi-execution-extrema.git
         </developerConnection>
         <tag>HEAD</tag>
     </scm>
