+++ release.properties
@@ -2,7 +2,7 @@
 #Fri Sep 30 12:01:14 IST 2016
 scm.tagNameFormat=@{project.artifactId}-@{project.version}
 pushChanges=true
-scm.url=scm\:git\:https\://github.com/wso2-extensions/siddhi-execution-timeseries.git
+scm.url=scm\:git\:https\://github.com/wso2-extensions/siddhi-execution-extrema.git
 preparationGoals=clean install
 scm.id=my-scm-server
 remoteTagging=true
